feat(mapi): factored out a common function into a separate file
refs. metriport/metriport#960
