+++ packages/lambdas/src/document-downloader.ts
@@ -13,6 +13,7 @@ import AWS from "aws-sdk";
 import { capture } from "./shared/capture";
 import { getEnv, getEnvOrFail } from "./shared/env";
 import { S3Utils } from "./shared/s3";
+import { getFileInfoFromS3 } from "./shared/file-info";
 
 // Keep this as early on the file as possible
 capture.init();
@@ -263,25 +264,25 @@ function oid(id: string): string {
   return `${OID_PREFIX}${id}`;
 }
 
-export async function getFileInfoFromS3(
-  key: string,
-  bucket: string
-): Promise<
-  | { exists: true; size: number; contentType: string }
-  | { exists: false; size?: never; contentType?: never }
-> {
-  try {
-    const head = await s3client
-      .headObject({
-        Bucket: bucket,
-        Key: key,
-      })
-      .promise();
-    return { exists: true, size: head.ContentLength ?? 0, contentType: head.ContentType ?? "" };
-  } catch (err) {
-    return { exists: false };
-  }
-}
+// export async function getFileInfoFromS3(
+//   key: string,
+//   bucket: string
+// ): Promise<
+//   | { exists: true; size: number; contentType: string }
+//   | { exists: false; size?: never; contentType?: never }
+// > {
+//   try {
+//     const head = await s3client
+//       .headObject({
+//         Bucket: bucket,
+//         Key: key,
+//       })
+//       .promise();
+//     return { exists: true, size: head.ContentLength ?? 0, contentType: head.ContentType ?? "" };
+//   } catch (err) {
+//     return { exists: false };
+//   }
+// }
 
 export function removeAndReturnB64FromXML(htmlString: string): { newXML: string; b64: string } {
   const openingTag = "<text";
